requirements.txt
cycler==0.10.0
decorator==4.4.0
Django==2.2.4
django-cors-headers==3.1.0
django-rest-framework==0.1.0
djangorestframework==3.10.2
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
pyparsing==2.4.2
pytesseract==0.2.9
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
